refactor(index): update home page
* Pass file to typescript
* Use custom hook instead of useEffect
* Switch to components instead of @mui
